supabase/functions/easyocr-service/requirements.txt
easyocr
pdf2image
Pillow
numpy
opencv-python-headless

